Move general error reporting (super)classes to root KGX package python file.
