+++ kgx/graph_operations/meta_knowledge_graph.py
@@ -8,11 +8,10 @@ import yaml
 from json import dump
 from json.encoder import JSONEncoder
 
-from kgx import ErrorDetecting
+from kgx import ErrorDetecting, ErrorType, MessageLevel, ValidationError
 from kgx.utils.kgx_utils import GraphEntityType
 from kgx.prefix_manager import PrefixManager
 from kgx.graph.base_graph import BaseGraph
-from kgx.validator import ValidationError, MessageLevel, ErrorType
 
 """
 Generate a knowledge map that corresponds to TRAPI KnowledgeMap.
