+++ kgx/graph_operations/summarize_graph.py
@@ -9,11 +9,10 @@ import yaml
 from json import dump
 from json.encoder import JSONEncoder
 
-from kgx import ErrorDetecting
+from kgx import ErrorDetecting, ErrorType, MessageLevel
 from kgx.utils.kgx_utils import GraphEntityType
 from kgx.graph.base_graph import BaseGraph
 from kgx.prefix_manager import PrefixManager
-from kgx.validator import MessageLevel, ErrorType
 
 TOTAL_NODES = "total_nodes"
 NODE_CATEGORIES = "node_categories"
